+++ backend/Core/Entities/CollatzConjecture.cs
@@ -2,8 +2,8 @@
 
 public class CollatzConjecture : BaseEntity
 {
-    public int StartingNumber { get; set; }
-    public List<int> Sequence { get; set; }
+    public long StartingNumber { get; set; }
+    public List<long> Sequence { get; set; }
     public int NumSteps { get; set; }
     public DateTime Timestamp { get; set; }
 }
